Removed mocks and Gmock
