+++ COMP72070-Section1-Group1.sln
@@ -13,8 +13,6 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Tamagotchi", "Tamagotchi\Ta
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ClientCLI", "ClientCLI\ClientCLI.vcxproj", "{BF6D05A8-9844-4EFB-878D-28EF044E42B0}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UnitTests", "UnitTests\UnitTests.vcxproj", "{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Logger", "Logger\Logger.vcxproj", "{439D3B44-6DA4-4805-860E-B5D2EEB5BCAB}"
 EndProject
 Global
@@ -65,14 +63,6 @@ Global
 		{BF6D05A8-9844-4EFB-878D-28EF044E42B0}.Release|x64.Build.0 = Release|x64
 		{BF6D05A8-9844-4EFB-878D-28EF044E42B0}.Release|x86.ActiveCfg = Release|Win32
 		{BF6D05A8-9844-4EFB-878D-28EF044E42B0}.Release|x86.Build.0 = Release|Win32
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Debug|x64.ActiveCfg = Debug|x64
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Debug|x64.Build.0 = Debug|x64
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Debug|x86.ActiveCfg = Debug|Win32
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Debug|x86.Build.0 = Debug|Win32
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Release|x64.ActiveCfg = Release|x64
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Release|x64.Build.0 = Release|x64
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Release|x86.ActiveCfg = Release|Win32
-		{46F6A5C6-82E0-43EC-8C2E-5483BE1B9505}.Release|x86.Build.0 = Release|Win32
 		{439D3B44-6DA4-4805-860E-B5D2EEB5BCAB}.Debug|x64.ActiveCfg = Debug|x64
 		{439D3B44-6DA4-4805-860E-B5D2EEB5BCAB}.Debug|x64.Build.0 = Debug|x64
 		{439D3B44-6DA4-4805-860E-B5D2EEB5BCAB}.Debug|x86.ActiveCfg = Debug|Win32
