+++ Client/Client.h
@@ -2,7 +2,6 @@
 #include <string>
 #include <memory>
 #include "../Communication/Communication.h"
-#include "gmock/gmock.h"
 
 namespace Client {
 	__interface IClient
@@ -18,12 +17,5 @@ namespace Client {
 		Client();
 		Client(std::unique_ptr<Communicators::ISender> sender);
 		std::unique_ptr<Data::IServerResponse> SendCommand(std::unique_ptr < Data::IClientRequest> request);
-	};
-}
-
-namespace Mocks {
-	class ClientMock : public Client::IClient {
-	public:
-		MOCK_METHOD(std::unique_ptr<Data::IServerResponse>, SendCommand, (std::unique_ptr<Data::IClientRequest>));
 	};
 }
