--- Communication/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="gmock" version="1.11.0" targetFramework="native" />
-</packages>
