+++ Logger/Logger.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <ctime>
 #include <vector>
-#include "gmock/gmock.h"
 
 namespace Logger {
 	enum action {
@@ -27,14 +26,5 @@ namespace Logger {
 		std::time_t getTime();
 		const std::vector<uint8_t> getData();
 		action getAction();
-	};
-}
-
-namespace Mocks {
-	class LogMock : public Logger::ILog {
-	public:
-		MOCK_METHOD(time_t, getTime, ());
-		MOCK_METHOD(const std::vector<uint8_t>, getData, ());
-		MOCK_METHOD(Logger::action, getAction, ());
 	};
 }
