--- UnitTests/UnitTests.cpp
@@ -1,15 +1,0 @@
-#include "pch.h"
-#include "CppUnitTest.h"
-#include "Communication.h"
-#include "gmock/gmock.h"
-#include <string>
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-namespace UnitTests
-{
-	TEST_CLASS(Data_TESTS)
-	{
-	public:
-	};
-}
